refactor(ops): extract OrderBy/Limit and Mutate/Unnest ops
Two more pairs of small "transparent" Op classes pulled out:

- _order_limit.py (122 lines): SemanticOrderByOp + SemanticLimitOp.
  Both delegate every metadata accessor to self.source and only
  override to_untagged() and __repr__.
- _mutate_unnest.py (167 lines): SemanticMutateOp + SemanticUnnestOp.
  Same pass-through pattern; struct unpacking lives in Unnest's
  to_untagged.

Helpers used inside to_untagged (_to_untagged, _resolve_expr, _unwrap)
are imported lazily to keep the modules independent of _core import order.

_core.py now 3,551 lines. 930 tests pass.
