+++ src/boring_semantic_layer/ops/_core.py
@@ -2644,141 +2644,7 @@ class SemanticAggregateOp(Relation):
         return out.mutate(**calc_cols)
 
 
-class SemanticMutateOp(Relation):
-    source: Relation
-    post: dict[
-        str,
-        Callable,
-    ]  # Transformed to FrozenDict[str, _CallableWrapper] in __init__
-    nested_columns: tuple[
-        str,
-        ...,
-    ] = ()  # Inherited from source if it has nested columns
-
-    def __init__(
-        self,
-        source: Relation,
-        post: dict[str, Callable] | None,
-        nested_columns: tuple[str, ...] = (),
-    ) -> None:
-        frozen_post = FrozenDict(
-            {name: _ensure_wrapped(fn) for name, fn in (post or {}).items()},
-        )
-        source_nested = nested_columns if nested_columns else getattr(source, "nested_columns", ())
-
-        super().__init__(
-            source=Relation.__coerce__(source),
-            post=frozen_post,
-            nested_columns=source_nested,
-        )
-
-    def __repr__(self) -> str:
-        return _semantic_repr(self)
-
-    @property
-    def values(self) -> FrozenOrderedDict[str, Any]:
-        return self.source.values
-
-    @property
-    def schema(self) -> Schema:
-        return self.source.schema
-
-    def to_untagged(self):
-        agg_tbl = _to_untagged(self.source)
-
-        # Process mutations incrementally so each can reference previous ones
-        # This allows: .mutate(rank=..., is_other=lambda t: t["rank"] > 5)
-        current_tbl = agg_tbl
-        for name, fn_wrapped in self.post.items():
-            proxy = MeasureScope(_tbl=current_tbl, _known=[], _post_agg=True)
-            resolved = _resolve_expr(_unwrap(fn_wrapped), proxy)
-
-            new_col = resolved.name(name)
-            current_tbl = current_tbl.mutate([new_col])
-
-        return current_tbl
-
-    def get_dimensions(self) -> Mapping[str, Dimension]:
-        """Get dictionary of dimensions from source."""
-        return self.source.get_dimensions()
-
-    def get_measures(self) -> Mapping[str, Measure]:
-        """Get dictionary of measures from source."""
-        return self.source.get_measures()
-
-    def get_calculated_measures(self) -> Mapping[str, Any]:
-        """Get dictionary of calculated measures from source."""
-        return self.source.get_calculated_measures()
-
-
-class SemanticUnnestOp(Relation):
-    """Unnest an array column, expanding rows (like Malloy's nested data pattern)."""
-
-    source: Relation
-    column: str
-
-    def __repr__(self) -> str:
-        return _semantic_repr(self)
-
-    @property
-    def schema(self) -> Schema:
-        # After unnesting, the schema changes - the array column is replaced by its element schema
-        # For now, delegate to source schema (ideally we'd update it)
-        return self.source.schema
-
-    @property
-    def values(self) -> FrozenDict:
-        return FrozenDict({})
-
-    def to_untagged(self):
-        """Convert to Ibis expression with functional struct unpacking.
-
-        Uses pure helper functions to extract struct fields when unnesting
-        produces struct columns that need to be expanded.
-        """
-
-        def build_struct_fields(col_expr, col_type):
-            """Pure function: build dict of struct field selections."""
-            return {name: col_expr[name] for name in col_type.names}
-
-        def unpack_struct_if_needed(unnested_tbl, column_name):
-            """Conditionally unpack struct fields into top-level columns."""
-            if column_name not in unnested_tbl.columns:
-                return unnested_tbl
-
-            col_expr = unnested_tbl[column_name]
-            col_type = col_expr.type()
-
-            # Only Struct types have fields to unpack
-            if isinstance(col_type, dt.Struct) and col_type.fields:
-                struct_fields = build_struct_fields(col_expr, col_type)
-                return unnested_tbl.select(unnested_tbl, **struct_fields)
-
-            return unnested_tbl
-
-        tbl = _to_untagged(self.source)
-
-        if self.column not in tbl.columns:
-            raise ValueError(f"Column '{self.column}' not found in table")
-
-        try:
-            unnested = tbl.unnest(self.column)
-        except Exception as e:
-            raise ValueError(f"Failed to unnest column '{self.column}': {e}") from e
-
-        return unpack_struct_if_needed(unnested, self.column)
-
-    def get_dimensions(self) -> Mapping[str, Dimension]:
-        """Get dictionary of dimensions from source."""
-        return self.source.get_dimensions()
-
-    def get_measures(self) -> Mapping[str, Measure]:
-        """Get dictionary of measures from source."""
-        return self.source.get_measures()
-
-    def get_calculated_measures(self) -> Mapping[str, Any]:
-        """Get dictionary of calculated measures from source."""
-        return self.source.get_calculated_measures()
+from ._mutate_unnest import SemanticMutateOp, SemanticUnnestOp  # noqa: E402, F401
 
 
 class SemanticJoinOp(Relation):
@@ -3642,97 +3508,7 @@ class SemanticJoinOp(Relation):
         )
 
 
-class SemanticOrderByOp(Relation):
-    source: Relation
-    keys: tuple[
-        str | ir.Value | Callable,
-        ...,
-    ]  # Transformed to tuple[str | _CallableWrapper, ...] in __init__
-
-    def __init__(self, source: Relation, keys: Iterable[str | ir.Value | Callable]) -> None:
-        def wrap_key(k):
-            return k if isinstance(k, str | _CallableWrapper) else _ensure_wrapped(k)
-
-        super().__init__(
-            source=Relation.__coerce__(source),
-            keys=tuple(wrap_key(k) for k in keys),
-        )
-
-    def __repr__(self) -> str:
-        return _semantic_repr(self)
-
-    @property
-    def values(self) -> FrozenOrderedDict[str, Any]:
-        return self.source.values
-
-    @property
-    def schema(self) -> Schema:
-        return self.source.schema
-
-    def to_untagged(self):
-        tbl = _to_untagged(self.source)
-
-        def resolve_order_key(key):
-            if isinstance(key, str):
-                return tbl[key] if key in tbl.columns else getattr(tbl, key, key)
-            elif isinstance(key, _CallableWrapper):
-                unwrapped = _unwrap(key)
-                return _resolve_expr(unwrapped, tbl)
-            return key
-
-        return tbl.order_by([resolve_order_key(key) for key in self.keys])
-
-    def get_dimensions(self) -> Mapping[str, Dimension]:
-        """Get dictionary of dimensions from source."""
-        return self.source.get_dimensions()
-
-    def get_measures(self) -> Mapping[str, Measure]:
-        """Get dictionary of measures from source."""
-        return self.source.get_measures()
-
-    def get_calculated_measures(self) -> Mapping[str, Any]:
-        """Get dictionary of calculated measures from source."""
-        return self.source.get_calculated_measures()
-
-
-class SemanticLimitOp(Relation):
-    source: Relation
-    n: int
-    offset: int
-
-    def __init__(self, source: Relation, n: int, offset: int = 0) -> None:
-        if n <= 0:
-            raise ValueError(f"limit must be positive, got {n}")
-        if offset < 0:
-            raise ValueError(f"offset must be non-negative, got {offset}")
-        super().__init__(source=Relation.__coerce__(source), n=n, offset=offset)
-
-    def __repr__(self) -> str:
-        return _semantic_repr(self)
-
-    @property
-    def values(self) -> FrozenOrderedDict[str, Any]:
-        return self.source.values
-
-    @property
-    def schema(self) -> Schema:
-        return self.source.schema
-
-    def to_untagged(self):
-        tbl = _to_untagged(self.source)
-        return tbl.limit(self.n) if self.offset == 0 else tbl.limit(self.n, offset=self.offset)
-
-    def get_dimensions(self) -> Mapping[str, Dimension]:
-        """Get dictionary of dimensions from source."""
-        return self.source.get_dimensions()
-
-    def get_measures(self) -> Mapping[str, Measure]:
-        """Get dictionary of measures from source."""
-        return self.source.get_measures()
-
-    def get_calculated_measures(self) -> Mapping[str, Any]:
-        """Get dictionary of calculated measures from source."""
-        return self.source.get_calculated_measures()
+from ._order_limit import SemanticLimitOp, SemanticOrderByOp  # noqa: E402, F401
 
 
 from ._index_op import (  # noqa: E402
