+++ src/boring_semantic_layer/ops/_order_limit.py
@@ -1,0 +1,118 @@
+"""``SemanticOrderByOp`` and ``SemanticLimitOp`` — terminal pass-through ops.
+
+Both are thin wrappers that delegate dimension/measure metadata to their
+source and only modify ``to_untagged()`` output. They share the same
+shape: pass-through ``schema``/``values``, override ``__repr__`` and
+``to_untagged``, and forward ``get_*`` accessors to ``self.source``.
+"""
+
+from __future__ import annotations
+
+from collections.abc import Callable, Iterable, Mapping
+from typing import Any
+
+from ibis.expr import types as ir
+from ibis.expr.operations.relations import Relation
+from ibis.expr.schema import Schema
+
+from .._xorq import FrozenOrderedDict
+from ._callable import _CallableWrapper, _ensure_wrapped
+from ._format import _semantic_repr
+from ._values import Dimension, Measure
+
+
+class SemanticOrderByOp(Relation):
+    source: Relation
+    keys: tuple[
+        str | ir.Value | Callable,
+        ...,
+    ]  # Transformed to tuple[str | _CallableWrapper, ...] in __init__
+
+    def __init__(self, source: Relation, keys: Iterable[str | ir.Value | Callable]) -> None:
+        def wrap_key(k):
+            return k if isinstance(k, str | _CallableWrapper) else _ensure_wrapped(k)
+
+        super().__init__(
+            source=Relation.__coerce__(source),
+            keys=tuple(wrap_key(k) for k in keys),
+        )
+
+    def __repr__(self) -> str:
+        return _semantic_repr(self)
+
+    @property
+    def values(self) -> FrozenOrderedDict[str, Any]:
+        return self.source.values
+
+    @property
+    def schema(self) -> Schema:
+        return self.source.schema
+
+    def to_untagged(self):
+        from ._core import _resolve_expr, _to_untagged, _unwrap
+
+        tbl = _to_untagged(self.source)
+
+        def resolve_order_key(key):
+            if isinstance(key, str):
+                return tbl[key] if key in tbl.columns else getattr(tbl, key, key)
+            elif isinstance(key, _CallableWrapper):
+                unwrapped = _unwrap(key)
+                return _resolve_expr(unwrapped, tbl)
+            return key
+
+        return tbl.order_by([resolve_order_key(key) for key in self.keys])
+
+    def get_dimensions(self) -> Mapping[str, Dimension]:
+        """Get dictionary of dimensions from source."""
+        return self.source.get_dimensions()
+
+    def get_measures(self) -> Mapping[str, Measure]:
+        """Get dictionary of measures from source."""
+        return self.source.get_measures()
+
+    def get_calculated_measures(self) -> Mapping[str, Any]:
+        """Get dictionary of calculated measures from source."""
+        return self.source.get_calculated_measures()
+
+
+class SemanticLimitOp(Relation):
+    source: Relation
+    n: int
+    offset: int
+
+    def __init__(self, source: Relation, n: int, offset: int = 0) -> None:
+        if n <= 0:
+            raise ValueError(f"limit must be positive, got {n}")
+        if offset < 0:
+            raise ValueError(f"offset must be non-negative, got {offset}")
+        super().__init__(source=Relation.__coerce__(source), n=n, offset=offset)
+
+    def __repr__(self) -> str:
+        return _semantic_repr(self)
+
+    @property
+    def values(self) -> FrozenOrderedDict[str, Any]:
+        return self.source.values
+
+    @property
+    def schema(self) -> Schema:
+        return self.source.schema
+
+    def to_untagged(self):
+        from ._core import _to_untagged
+
+        tbl = _to_untagged(self.source)
+        return tbl.limit(self.n) if self.offset == 0 else tbl.limit(self.n, offset=self.offset)
+
+    def get_dimensions(self) -> Mapping[str, Dimension]:
+        """Get dictionary of dimensions from source."""
+        return self.source.get_dimensions()
+
+    def get_measures(self) -> Mapping[str, Measure]:
+        """Get dictionary of measures from source."""
+        return self.source.get_measures()
+
+    def get_calculated_measures(self) -> Mapping[str, Any]:
+        """Get dictionary of calculated measures from source."""
+        return self.source.get_calculated_measures()
